Do more refactor
- Move Filemanager from Remote to its separate file
- Avoid doing third party work at FileManager.copyFromTo method, move
it to a block.
- Add tldr related file operations in an extension of Filemanager
