+++ TLDR/AppDelegate.swift
@@ -14,7 +14,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     var window: UIWindow?
 
     func application(application: UIApplication, didFinishLaunchingWithOptions launchOptions: [NSObject: AnyObject]?) -> Bool {
-        FileManager.copyBundleToDocument(false)
+        FileManager.copyBundleToDocument(replace: false, onSuccess: {
+            StoreManager.updateDB()
+        })
         return true
     }
